// src/types/partials/DocumentReferenceContext.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TPeriod } from '../partials/Period';

export type TDocumentReferenceContext = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    encounter?: TReference[];
    event?: TCodeableConcept[];
    period?: TPeriod;
    facilityType?: TCodeableConcept;
    practiceSetting?: TCodeableConcept;
    sourcePatientInfo?: TReference;
    related?: TReference[];
};